requirements.txt
Flask==2.0.3
tensorflow==2.9.1  # Or the version you're using
Pillow==9.1.0
